web/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=44.0.0
litellm>=1.50.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
alembic>=1.14.0
loguru>=0.7.0
